Update UI and tests
