+++ ui-tests/playwright.base.config.ts
@@ -4,7 +4,9 @@ export default defineConfig({
     testDir: 'tests/e2e',
     workers: 1,
     forbidOnly: !!process.env.CI, // Fail the build on CI if you accidentally left test.only in the source code
-    retries: 2,
+    retries: 1,
+    // Stop the run gracefully before the CI job's 30-minute step timeout so the HTML report is still written and uploaded.
+    globalTimeout: 25 * 60 * 1000,
     reporter: [['html', { open: 'on-failure', host: '0.0.0.0', port: 9323 }]],
     use: {
         timezoneId: 'Europe/London',
