+++ ui-tests/tests/e2e/saved-graphs.spec.ts
@@ -1,14 +1,15 @@
-import { expect, Page, test } from '@playwright/test';
+import { expect, test } from '@playwright/test';
 import {
-    clickSavedGraphsFolder,
     clickSavedGraphsGraph,
-    waitForLayoutToFinish,
-} from './utils';
+    navigateInSavedGraphs,
+    OPEN_SAVED_GRAPH_METHODS,
+} from './saved-graphs.utils';
+import { waitForLayoutToFinish } from './utils';
 
 const PAGE_SIZE = 8;
 
 test('Saved graphs table is visible', async ({ page }) => {
-    await navigateToSavedGraphsFolder(page, 'vanilla');
+    await navigateInSavedGraphs(page, { namespace: 'vanilla' });
     await clickSavedGraphsGraph(page, 'event');
     await page.getByRole('button', { name: 'event GRAPH' }).click();
     await expect(
@@ -19,7 +20,7 @@ test('Saved graphs table is visible', async ({ page }) => {
 });
 
 test(`Card view has N cards per page`, async ({ page }) => {
-    await navigateToSavedGraphsFolder(page, 'vanilla');
+    await navigateInSavedGraphs(page, { namespace: 'vanilla' });
     await expect(page.getByText(`1-${PAGE_SIZE} of`)).toBeVisible();
     await expect(page.getByRole('button', { name: 'GRAPH' })).toHaveCount(
         PAGE_SIZE,
@@ -27,7 +28,11 @@ test(`Card view has N cards per page`, async ({ page }) => {
 });
 
 test('Row sorting on saved graphs table by columns', async ({ page }) => {
-    await navigateToSavedGraphsFolder(page, 'vanilla', 'table');
+    await navigateInSavedGraphs(
+        page,
+        { namespace: 'vanilla' },
+        { view: 'table' },
+    );
     const table = await page.getByRole('table');
 
     // Name
@@ -98,54 +103,29 @@ test('Row sorting on saved graphs table by columns', async ({ page }) => {
     await expect(firstRowEdgeCountDescendingName).toBe('0');
 });
 
-test('Open graph by double clicking', async ({ page }) => {
-    await navigateToSavedGraphsFolder(page, 'vanilla', 'table');
-    await page
-        .getByRole('cell', { name: 'temporal_props', exact: true })
-        .dblclick();
-    await expect(page).toHaveURL(
-        /\/graph\/vanilla\/temporal_props\?initialNodes=%5B%5D/,
-    );
-});
-
-test('Open graph by clicking open button on rhs panel', async ({ page }) => {
-    await navigateToSavedGraphsFolder(page, 'vanilla', 'table');
-    await page
-        .getByRole('cell', { name: 'temporal_props', exact: true })
-        .click();
-    await page
-        .getByRole('heading', { name: 'vanilla/temporal_props' })
-        .isVisible();
-    await page.getByText('Namespace').isVisible();
-    await page
-        .getByLabel('Button group with a nested menu')
-        .getByRole('button', { name: 'Open' })
-        .click();
-
-    await page.waitForSelector('text=Overview');
-    await page.getByText('Overview').isVisible();
-    await expect(page).toHaveURL(
-        /\/graph\/vanilla\/temporal_props\?initialNodes=%5B%5D/,
-    );
-});
-
-test('Check properties visible on right hand side and open graph from minimap preview', async ({
-    page,
-}) => {
-    await navigateToSavedGraphsFolder(page, 'vanilla');
-    await clickSavedGraphsGraph(page, 'event');
-    await expect(page.getByText('Namespace')).toBeVisible();
-    await page.getByRole('button', { name: 'Open' }).nth(1).click();
-    await waitForLayoutToFinish(page);
-    await expect(page).toHaveURL(
-        /\/graph\/vanilla\/event\?initialNodes=%5B%5D/,
-    );
+test('Open graph by all available methods', async ({ page }) => {
+    for (const method of OPEN_SAVED_GRAPH_METHODS) {
+        await navigateInSavedGraphs(
+            page,
+            { namespace: 'vanilla', graphName: 'temporal_props' },
+            {
+                method,
+            },
+        );
+        await expect(page).toHaveURL(
+            /\/graph\/vanilla\/temporal_props\?initialNodes=%5B%5D/,
+        );
+    }
 });
 
 test('Search saved graphs table, clear search and hide search', async ({
     page,
 }) => {
-    await navigateToSavedGraphsFolder(page, 'vanilla', 'table');
+    await navigateInSavedGraphs(
+        page,
+        { namespace: 'vanilla' },
+        { view: 'table' },
+    );
     const table = await page.getByRole('table');
     await page.getByRole('button', { name: 'Show/Hide search' }).click();
     const searchInput = page.getByRole('textbox', {
@@ -162,7 +142,11 @@ test('Search saved graphs table, clear search and hide search', async ({
 });
 
 test('Filter by Columns', async ({ page }) => {
-    await navigateToSavedGraphsFolder(page, 'vanilla', 'table');
+    await navigateInSavedGraphs(
+        page,
+        { namespace: 'vanilla' },
+        { view: 'table' },
+    );
     const table = await page.getByRole('table');
     await page.getByRole('button', { name: 'Show/Hide filters' }).click();
     const filterNameInput = page.getByPlaceholder('Filter by Name');
@@ -186,38 +170,32 @@ test('Filter by Columns', async ({ page }) => {
 });
 
 test('Switching between previews', async ({ page }) => {
-    await navigateToSavedGraphsFolder(page, 'vanilla');
+    await navigateInSavedGraphs(page, { namespace: 'vanilla' });
     await page.getByRole('button', { name: 'Expand details' }).click();
     await page.waitForTimeout(500);
     await clickSavedGraphsGraph(page, 'event');
     await waitForLayoutToFinish(page);
     expect(
-        await page.getByRole('button', { name: 'Open' }).nth(1).screenshot(),
+        await page
+            .getByRole('region', { name: 'Graph preview' })
+            .getByRole('button', { name: 'Open' })
+            .screenshot(),
     ).toMatchSnapshot('event-preview-first-click.png');
     await clickSavedGraphsGraph(page, 'persistent');
     await waitForLayoutToFinish(page);
     expect(
-        await page.getByRole('button', { name: 'Open' }).nth(1).screenshot(),
+        await page
+            .getByRole('region', { name: 'Graph preview' })
+            .getByRole('button', { name: 'Open' })
+            .screenshot(),
     ).toMatchSnapshot('persistent-preview-first-click.png');
     await clickSavedGraphsGraph(page, 'event');
     await waitForLayoutToFinish(page);
     // We expect no difference between the first time we preview and the second time
     expect(
-        await page.getByRole('button', { name: 'Open' }).nth(1).screenshot(),
+        await page
+            .getByRole('region', { name: 'Graph preview' })
+            .getByRole('button', { name: 'Open' })
+            .screenshot(),
     ).toMatchSnapshot('event-preview-first-click.png');
 });
-
-async function navigateToSavedGraphsFolder(
-    page: Page,
-    folder: string,
-    view?: 'card' | 'table',
-) {
-    await page.goto('/saved-graphs');
-    await clickSavedGraphsFolder(page, folder);
-    await page.waitForLoadState('networkidle');
-    if (view === 'table') {
-        await page
-            .getByRole('button', { name: 'Table view', exact: true })
-            .click();
-    }
-}
